+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 1.2.4
 
 * `connection_pool_max_size` and `max_connection_pool_count` are no longer supported due to issues with `ProcessPoolExecutor`, instead you may specify `get_http_pool` if you need to control these settings.
+* Add `file_size` argument to `BlobFile` to skip reading the size of the file on creation.
 
 ## 1.2.3